grpc-image-del-service/requirements.txt
grpcio-tools
pymongo
grpcio
protobuf
python-dotenv